requirements.txt
Django==4.2.5
django-cors-headers==4.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
python-dotenv==1.0.0
requests==2.31.0